move unit_types into wrapper
