+++ nidup/pysc2/information.py
@@ -1,7 +1,7 @@
 
 
 from nidup.pysc2.observations import Observations, ScreenFeatures
-from nidup.pysc2.unit_types import UnitTypeIds
+from nidup.pysc2.wrapper.unit_types import UnitTypeIds
 
 # Parameters
 _PLAYER_SELF = 1
