+++ nidup/pysc2/scripted_agents.py
@@ -7,7 +7,7 @@ import random
 import time
 
 from nidup.pysc2.actions import TerranActions, TerranActionIds
-from nidup.pysc2.unit_types import UnitTypeIds
+from nidup.pysc2.wrapper.unit_types import UnitTypeIds
 from nidup.pysc2.learning.game_results import GameResultsTable
